rename the `batch` module to `bank_erosion`
